refactor: move observed workflow helper to utils
