+++ src/commands/lint.rs
@@ -14,13 +14,11 @@ use homeboy::core::refactor::plan::{
     collect_refactor_sources, lint_refactor_request, LintSourceOptions,
 };
 
-use super::source_command::{
-    finish_observed_workflow, resolve_ci_job_for_command, resolve_source_context,
-    ObservedWorkflowRunner,
-};
+use super::source_command::{resolve_ci_job_for_command, resolve_source_context};
 use super::utils::args::{
     BaselineArgs, ExtensionOverrideArgs, HiddenJsonArgs, PositionalComponentArgs, SettingArgs,
 };
+use super::utils::observed_workflow::{finish_observed_workflow, ObservedWorkflowRunner};
 use super::{CmdResult, GlobalArgs};
 
 #[derive(Args)]
