+++ src/commands/utils/mod.rs
@@ -1,5 +1,6 @@
 pub mod args;
 pub mod entity_suggest;
+pub(crate) mod observed_workflow;
 pub mod resolve;
 pub mod resource_policy;
 pub mod response;
